impl user update

+++ router/api/v1/user/delete.go
@@ -14,9 +14,16 @@ func Delete() gin.HandlerFunc {
 			Abilities: []model.Ability{model.AbilityDeleteUser},
 			WhenYes: func(id string) {
 				if user := store.Access.GetUserById(ctx.Param("user_id")); user != nil {
+
+					if store.Access.DeleteUser(user.Id) != nil {
+						ctx.JSON(http.StatusForbidden, gin.H{
+							"message": "cant_delete_account",
+						})
+						return
+					}
+
 					store.Access.DeleteSessionUserByUser(user)
 					store.Access.LeaveRoleUser(user.Id)
-					store.Access.DeleteUser(user.Id)
 					return
 				}
 				ctx.JSON(http.StatusNotFound, gin.H{
